1.2025.04.24: output: filters: correct colors

+++ src/imageproc/AdjustBrightness.cpp
@@ -16,17 +16,19 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "AdjustBrightness.h"
-#include <QImage>
 #include <stdexcept>
 #include <stdint.h>
+#include <QImage>
+#include "AdjustBrightness.h"
 
 namespace imageproc
 {
 
 void adjustBrightness(
-    QImage& rgb_image, QImage const& brightness,
-    double const wr, double const wb)
+    QImage& rgb_image,
+    QImage const& brightness,
+    float const wr,
+    float const wb)
 {
     switch (rgb_image.format())
     {
@@ -57,35 +59,35 @@ void adjustBrightness(
     int const width = rgb_image.width();
     int const height = rgb_image.height();
 
-    double const wg = 1.0 - wr - wb;
-    double const wu = (1.0 - wb);
-    double const wv = (1.0 - wr);
-    double const r_wg = 1.0 / wg;
-    double const r_wu = 1.0 / wu;
-    double const r_wv = 1.0 / wv;
+    float const wg = 1.0f - wr - wb;
+    float const wu = (1.0f - wb);
+    float const wv = (1.0f - wr);
+    float const r_wg = 1.0f / wg;
+    float const r_wu = 1.0f / wu;
+    float const r_wv = 1.0f / wv;
 
     for (int y = 0; y < height; ++y)
     {
         for (int x = 0; x < width; ++x)
         {
             uint32_t RGB = rgb_line[x];
-            double const R = (RGB >> 16) & 0xFF;
-            double const G = (RGB >> 8) & 0xFF;
-            double const B = RGB & 0xFF;
+            float const R = (RGB >> 16) & 0xFF;
+            float const G = (RGB >> 8) & 0xFF;
+            float const B = RGB & 0xFF;
 
-            double const Y = wr * R + wg * G + wb * B;
-            double const U = (B - Y) * r_wu;
-            double const V = (R - Y) * r_wv;
+            float const Y = wr * R + wg * G + wb * B;
+            float const U = (B - Y) * r_wu;
+            float const V = (R - Y) * r_wv;
 
-            double new_Y = br_line[x];
-            double new_R = new_Y + V * wv;
-            double new_B = new_Y + U * wu;
-            double new_G = (new_Y - new_R * wr - new_B * wb) * r_wg;
+            float new_Y = br_line[x];
+            float new_R = new_Y + V * wv;
+            float new_B = new_Y + U * wu;
+            float new_G = (new_Y - new_R * wr - new_B * wb) * r_wg;
 
             RGB &= 0xFF000000; // preserve alpha
-            RGB |= uint32_t(qBound(0, int(new_R + 0.5), 255)) << 16;
-            RGB |= uint32_t(qBound(0, int(new_G + 0.5), 255)) << 8;
-            RGB |= uint32_t(qBound(0, int(new_B + 0.5), 255));
+            RGB |= uint32_t(qBound(0, int(new_R + 0.5f), 255)) << 16;
+            RGB |= uint32_t(qBound(0, int(new_G + 0.5f), 255)) << 8;
+            RGB |= uint32_t(qBound(0, int(new_B + 0.5f), 255));
             rgb_line[x] = RGB;
         }
         rgb_line += rgb_wpl;
@@ -93,14 +95,18 @@ void adjustBrightness(
     }
 }
 
-void adjustBrightnessYUV(QImage& rgb_image, QImage const& brightness)
+void adjustBrightnessYUV(
+    QImage& rgb_image,
+    QImage const& brightness)
 {
-    adjustBrightness(rgb_image, brightness, 0.299, 0.114);
+    adjustBrightness(rgb_image, brightness, 0.299f, 0.114f);
 }
 
-void adjustBrightnessGrayscale(QImage& rgb_image, QImage const& brightness)
+void adjustBrightnessGrayscale(
+    QImage& rgb_image,
+    QImage const& brightness)
 {
-    adjustBrightness(rgb_image, brightness, 11.0/32.0, 5.0/32.0);
+    adjustBrightness(rgb_image, brightness, 11.0f/32.0f, 5.0f/32.0f);
 }
 
 } // namespace imageproc
